+++ docassemble_webapp/docassemble/webapp/microsoft.py
@@ -56,12 +56,14 @@ class azurekey(object):
         os.utime(filename, (secs, secs))
     def set_contents_from_filename(self, filename):
         self.azure_object.conn.create_blob_from_path(self.azure_object.container, self.name, filename)
-        properties = self.azure_object.conn.get_blob_properties(self.azure_object.container, self.name).properties
-        secs = (properties.last_modified - epoch).total_seconds()
+        self.get_properties()
+        secs = (self.last_modified - epoch).total_seconds()
         os.utime(filename, (secs, secs))
     def set_contents_from_string(self, text):
         self.azure_object.conn.create_blob_from_text(self.azure_object.container, self.name, text)
     def get_epoch_modtime(self):
+        if not hasattr(self, 'last_modified'):
+            self.get_properties()
         return (self.last_modified - epoch).total_seconds()
     def generate_url(self, seconds, display_filename=None, content_type=None):
         if content_type is None: