+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - PDF page retrieval index error.
 - File permissions issue in S3 that prevents server restart.  S3 users 
   must upgrade the Docker image.
-- Incorrect timestamps when downloading from S3
+- Incorrect timestamps when downloading from persistent cloud storage.
 
 ## [0.1.78] - 2017-10-22
 ### Changed